Feat: New validation for Registration and Event module. New getQrCode and Confirm-presence action

+++ backend/app/Services/EventService.php
@@ -5,6 +5,7 @@ namespace App\Services;
 use App\DTO\Event\EventDTO;
 use App\DTO\DTO;
 use App\Models\Event;
+use Exception;
 use Illuminate\Support\Collection;
 
 class EventService
@@ -29,19 +30,34 @@ class EventService
         $arr = $eventDto->toArray();
         foreach ($arr as $name => $value)
             if($value != null) $obj->$name = $value;
+        $obj->owner = auth()->id();
         return $obj->save();
     }
 
+    /**
+     * @throws Exception
+     */
     public function update(DTO $eventDto, int $id):bool{
         $obj = Event::findByOrFail($id);
         $arr = $eventDto->toArray();
+
+        if($obj->owner != auth()->id())
+            throw new Exception("You do not have privileges to perform this action", 401);
+
         foreach ($arr as $name => $value)
             if($value != null) $obj->$name = $value;
         return $obj->save();
     }
 
+    /**
+     * @throws Exception
+     */
     public function delete(int $id):bool{
         $obj = Event::findByOrFail($id);
+
+        if($obj->owner != auth()->id())
+            throw new Exception("You do not have privileges to perform this action", 401);
+
         return $obj->delete();
     }
